add clean cache test

+++ test_sh/rocksdb_test_4value.sh
@@ -1,6 +1,7 @@
 #! /bin/sh
 
-value_array=(1024 4096 16384 65536)
+#value_array=(1024 4096 16384 65536)
+value_array=(4096)
 test_all_size=81920000000   #80G
 
 
@@ -10,7 +11,8 @@ bench_compression="none" #"snappy,none"
 
 #bench_benchmarks="fillseq,stats,readseq,readrandom,stats" #"fillrandom,fillseq,readseq,readrandom,stats"
 #bench_benchmarks="fillrandom,stats,readseq,readrandom,readrandom,readrandom,stats"
-bench_benchmarks="fillrandom,stats,wait,stats,readseq,readrandom,readrandom,readrandom,stats"
+#bench_benchmarks="fillrandom,stats,wait,stats,readseq,readrandom,readrandom,readrandom,stats"
+bench_benchmarks="fillrandom,stats,wait,clean_cache,stats,readseq,readrandom,stats"
 #bench_benchmarks="fillseq,stats"
 bench_num="20000000"
 bench_readnum="1000000"
@@ -52,7 +54,9 @@ RUN_ONE_TEST() {
 }
 
 CLEAN_CACHE() {
-    rm -f $bench_db_path/*
+    if [ -n "$bench_db_path" ];then
+        rm -f $bench_db_path/*
+    fi
     sleep 2
     sync
     echo 3 > /proc/sys/vm/drop_caches
